+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.x
+
+## Bug fixes
+- PIM-4494: Fix loading page when family has been sorted
+
 # 1.3.16 (2015-06-08)
 
 ## Bug fixes